polyval: use `FieldElement` as common internal representation
Continued work towards consolidating and simplifying the implementation.

This uses a simple `[u8; 16]` as the internal representation,
effectively deferring the decoding step to SIMD registers, which has
been moved to a `From<FieldElement>` impl for both `__m128i` on
`x86`/`x86_64` and `uint8x16_t` on ARMv8. SIMD registers are still
loaded using the relevant intrinsics (`__mm_loadu_si128` and `vld1q_u8`
respectively).

With this change, all three backends: `x86`/`x86_64`, `armv8`, and
`soft`, all use the `FieldElement` type as their internal representation
for the `Polyval` type. This should make it possible to consolidate the
implementation into one that can be shared across backends.
